Common: Wifi: Remote Code Execution Vulnerability in Broadcom Wi-Fi Driver
CVE-2016-0801 ANDROID-26492805 (Broadcom ref#: RB#80670, RB#80821)
CVE-2016-2433 ANDROID-26523007 (Broadcom ref#: RB#80670, RB#80821)

This is a related issue with packet validation in the Broadcom wireless driver to #191524.
With this issue, there is a race condition between registering and deregistering a temporary interface that results in a kernel panic.
The fix is to filter out all packets type = ETHER_TYPE_BRCM(0x886c).

Bug: 28547399
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/229068
Tested-by: joe_tsai <joe_tsai@asus.com>
Reviewed-on: http://mcrd1-22-pc.corpnet.asus/code-review/master/229073

+++ drivers/net/wireless/bcmdhd/dhd_common.c
@@ -1644,7 +1644,7 @@ wl_show_host_event(dhd_pub_t *dhd_pub, wl_event_msg_t *event, void *event_data,
 #endif /* SHOW_EVENTS */
 
 int
-wl_host_event(dhd_pub_t *dhd_pub, int *ifidx, void *pktdata, uint16 pktlen,
+wl_host_event(dhd_pub_t *dhd_pub, int *ifidx, void *pktdata, size_t pktlen,
 	wl_event_msg_t *event, void **data_ptr, void *raw_event)
 {
 	/* check whether packet is a BRCM event pkt */
@@ -1652,7 +1652,7 @@ wl_host_event(dhd_pub_t *dhd_pub, int *ifidx, void *pktdata, uint16 pktlen,
 	uint8 *event_data;
 	uint32 type, status, datalen;
 	uint16 flags;
-	int evlen;
+	uint evlen;
 	int hostidx;
 
 	if (pktlen < sizeof(bcm_event_t))
